state goal agent passing tests

+++ universalSuccessorFeatures/agents/stateGoalAgent.py
@@ -150,8 +150,8 @@ class StateGoalAgent():
         with torch.no_grad():
             return torch.argmax(
                 self.policy_net(
-                    torch.tensor(agent_position).to(self.device),
-                    torch.tensor(goal_position).to(self.device)
+                    torch.tensor(agent_position).to(torch.float).to(self.device),
+                    torch.tensor(goal_position).to(torch.float).to(self.device)
                 )
             ).item()
